requirements.txt
openai-whisper
transformers
sentencepiece
pyqt5
sounddevice
numpy
